+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Acceptance-Rejection distribution.
 - Bates distribution.
 - Bernoulli distribution.
+- Binomial distribution.
 - Box-Muller distribution.
 - Irwin-Hall distribution.
 - Marsaglia distribution.